+++ test/helpers/web.test.ts
@@ -1,5 +1,9 @@
-import { HttpsProxyAgent } from 'https-proxy-agent'
-import nock from 'nock'
+import {
+  MockAgent,
+  MockClient,
+  ProxyAgent,
+  setGlobalDispatcher,
+} from 'undici'
 
 import { version } from '../../package.json'
 import {
@@ -22,6 +26,9 @@ describe('Web Helpers', () => {
   let uploadFile: string
   let query: string
   let source: string
+
+  let mockAgent: MockAgent
+  let mockClient: MockClient
   beforeEach(() => {
     token = '123-abc-890-xyz'
     uploadFile = 'some content'
@@ -29,7 +36,8 @@ describe('Web Helpers', () => {
     source = ''
 
     // deepcode ignore WrongNumberOfArgs/test: believe this is a false positive
-    nock('https://codecov.io').put('/').query(true).reply(200, 'testPUT')
+    mockAgent = new MockAgent({ connections: 1 })
+    setGlobalDispatcher(mockAgent)
   })
 
   afterEach(() => {
@@ -39,11 +47,11 @@ describe('Web Helpers', () => {
 
   it('Can POST to the uploader endpoint (HTTP)', async () => {
     uploadURL = 'http://codecov.io'
-    // deepcode ignore WrongNumberOfArgs/test: believe this is a false positive
-    nock('http://codecov.io')
-      .post('/upload/v4')
-      .query(true)
-      .reply(200, 'testPOSTHTTP')
+    mockClient = mockAgent.get(uploadURL)
+    mockClient.intercept({
+      method: 'POST',
+      path: `/upload/v4?package=uploader-${version}&token=${token}&hello`,
+    }).reply(200, 'testPOSTHTTP')
 
     const response = await uploadToCodecovPOST(
       new URL(uploadURL),
@@ -59,6 +67,7 @@ describe('Web Helpers', () => {
     try {
       expect(response).toBe('testPOSTHTTP')
     } catch (error) {
+      throw new Error(`${response} - ${error}`)
       console.trace(error)
     }
   })
@@ -66,11 +75,11 @@ describe('Web Helpers', () => {
   it('Can POST to the uploader endpoint (HTTPS)', async () => {
     jest.spyOn(console, 'log').mockImplementation(() => void {})
     uploadURL = 'https://codecov.io'
-    // deepcode ignore WrongNumberOfArgs/test: believe this is a false positive
-    nock('https://codecov.io')
-      .post('/upload/v4')
-      .query(true)
-      .reply(200, 'testPOSTHTTPS')
+    mockClient = mockAgent.get(uploadURL)
+    mockClient.intercept({
+      method: 'POST',
+      path: `/upload/v4?package=uploader-${version}&token=${token}&hello`,
+    }).reply(200, 'testPOSTHTTPS')
 
     const response = await uploadToCodecovPOST(
       new URL(uploadURL),
@@ -92,6 +101,13 @@ describe('Web Helpers', () => {
       putURL: new URL('https://codecov.io'),
       resultURL: new URL('https://results.codecov.io'),
     }
+
+    mockClient = mockAgent.get(postResults.putURL.origin)
+    mockClient.intercept({
+      method: 'PUT',
+      path: '/',
+    }).reply(200, postResults.resultURL.href)
+
     const response = await uploadToCodecovPUT(postResults, uploadFile, {
       flags: '',
       slug: '',
@@ -249,13 +265,13 @@ describe('Web Helpers', () => {
 
     it('will return an object when parsing correctly and input has multiple linebreaks', () => {
       const testURL = `https://result.codecov.local
-      
-      
-      
-      
-      
-      
-      
+
+
+
+
+
+
+
       https://put.codecov.local`
       const expectedResults = {
         putURL: 'https://put.codecov.local/',
@@ -293,7 +309,7 @@ describe('generateRequestHeadersPOST()', () => {
       )}&token=134&slug=testOrg/testUploader`,
     )
     expect(typeof requestHeaders.options.body).toEqual('undefined')
-    expect(typeof requestHeaders.options.agent).toEqual('undefined')
+    expect(typeof requestHeaders.agent).toEqual('undefined')
   })
 
   it('should return return the correct url when args.upstream is set', () => {
@@ -313,8 +329,8 @@ describe('generateRequestHeadersPOST()', () => {
     )
 
     expect(typeof requestHeaders.options.body).toEqual('undefined')
-    expect(requestHeaders.options.agent).toMatchObject(
-      new HttpsProxyAgent(args.upstream),
+    expect(requestHeaders.agent).toMatchObject(
+      new ProxyAgent(args.upstream),
     )
   })
 })
@@ -332,7 +348,7 @@ describe('generateRequestHeadersPUT()', () => {
 
     expect(requestHeaders.url.href).toEqual('https://localhost.local/')
     expect(requestHeaders.options.body).toEqual("I'm a coverage report!")
-    expect(typeof requestHeaders.options.agent).toEqual('undefined')
+    expect(typeof requestHeaders.agent).toEqual('undefined')
   })
 
   it('should return return the correct url when args.upstream is set', () => {
@@ -345,8 +361,8 @@ describe('generateRequestHeadersPUT()', () => {
 
     expect(requestHeaders.url.href).toEqual('https://localhost.local/')
     expect(requestHeaders.options.body).toEqual("I'm a coverage report!")
-    expect(requestHeaders.options.agent).toMatchObject(
-      new HttpsProxyAgent(args.upstream),
+    expect(requestHeaders.agent).toMatchObject(
+      new ProxyAgent(args.upstream),
     )
   })
 })