+++ test/index.test.ts
@@ -1,11 +1,19 @@
 import Module from 'module'
 
+import fs from 'fs'
+import td from 'testdouble'
+import {
+  MockAgent,
+  MockClient,
+  setGlobalDispatcher,
+} from 'undici'
+
 import * as app from '../src'
 
 import { version } from '../package.json'
-import nock from 'nock'
-import fs from 'fs'
 import { UploadLogger } from '../src/helpers/logger'
+import { detectProvider } from '../src/helpers/provider'
+import * as webHelpers from '../src/helpers/web'
 
 // Backup the env
 const realEnv = { ...process.env }
@@ -14,6 +22,9 @@ describe('Uploader Core', () => {
   const env = process.env
 
   UploadLogger.setLogLevel('verbose')
+
+  let mockAgent: MockAgent
+  let mockClient: MockClient
 
   beforeEach(() => {
     // https://bensmithgall.com/blog/jest-mock-trick if this works!
@@ -25,17 +36,20 @@ describe('Uploader Core', () => {
 
     jest.mock('process', () => mockProcess())
     jest.resetAllMocks()
+
+    mockAgent = new MockAgent({ connections: 1 })
+    setGlobalDispatcher(mockAgent)
   })
 
   afterEach(() => {
     process.env = env
     jest.restoreAllMocks()
+    td.reset()
   })
 
   it('Can return version', () => {
     expect(app.getVersion()).toBe(version)
   })
-
   it('Can display header', () => {
     expect(app.generateHeader(app.getVersion())).toBe(`
      _____          _
@@ -55,21 +69,31 @@ describe('Uploader Core', () => {
     process.env.CI = 'true'
     process.env.CIRCLECI = 'true'
 
-    nock('https://codecov.io')
-      .post('/upload/v4')
-      .query(parsedObj => parsedObj.name === 'customname')
-      .reply(200, 'https://results.codecov.io\nhttps://codecov.io')
-
-    nock('https://codecov.io').put('/').reply(200, 'success')
-
-    const result = await app.main({
-      name: 'customname',
-      token: 'abcdefg',
-      url: 'https://codecov.io',
+    const args = {
       flags: '',
+      name: 'customname',
       slug: '',
-      upstream: ''
-    })
+      token: 'abcdefg',
+      upstream: '',
+      url: 'https://codecov.io',
+    }
+    const inputs = { args, environment: process.env }
+    const serviceParams = detectProvider(inputs, args.token != '')
+    const buildParams = webHelpers.populateBuildParams(inputs, serviceParams)
+    const query = webHelpers.generateQuery(buildParams)
+
+    mockClient = mockAgent.get('https://codecov.io')
+    mockClient.intercept({
+      method: 'POST',
+      path: `/upload/v4?package=uploader-${version}&token=${args.token}&${query}`,
+    }).reply(200, 'https://results.codecov.io\nhttps://codecov.io')
+
+    mockClient.intercept({
+      method: 'PUT',
+      path: '/',
+    }).reply(200, 'success')
+
+    const result = await app.main(args)
     expect(result).toEqual({
       status: 'success',
       resultURL: 'https://results.codecov.io/',
@@ -133,14 +157,7 @@ describe('Uploader Core', () => {
       process.env.CI = 'true'
       process.env.CIRCLECI = 'true'
 
-      nock('https://codecov.io')
-        .post('/upload/v4')
-        .query(actualQueryObject => actualQueryObject.flags === 'a-flag')
-        .reply(200, 'https://results.codecov.io\nhttps://codecov.io')
-
-      nock('https://codecov.io').put('/').reply(200, 'success')
-
-      const result = await app.main({
+      const args = {
         flags: 'a-flag',
         token: 'abcdefg',
         url: 'https://codecov.io',
@@ -148,7 +165,24 @@ describe('Uploader Core', () => {
         source: '',
         slug: '',
         upstream: ''
-      })
+      }
+      const inputs = { args, environment: process.env }
+      const serviceParams = detectProvider(inputs, args.token != '')
+      const buildParams = webHelpers.populateBuildParams(inputs, serviceParams)
+      const query = webHelpers.generateQuery(buildParams)
+
+      mockClient = mockAgent.get('https://codecov.io')
+      mockClient.intercept({
+        method: 'POST',
+        path: `/upload/v4?package=uploader-${version}&token=${args.token}&${query}`,
+      }).reply(200, 'https://results.codecov.io\nhttps://codecov.io')
+
+      mockClient.intercept({
+        method: 'PUT',
+        path: '/',
+      }).reply(200, 'success')
+
+      const result = await app.main(args)
       expect(result).toEqual({
         status: 'success',
         resultURL: 'https://results.codecov.io/',
@@ -161,22 +195,31 @@ describe('Uploader Core', () => {
     process.env.CIRCLECI = 'true'
 
     const parent = '2x4bqz123abc'
-
-    nock('https://codecov.io')
-      .post('/upload/v4')
-      .query(actualQueryObject => actualQueryObject.parent === parent)
-      .reply(200, 'https://results.codecov.io\nhttps://codecov.io')
-
-    nock('https://codecov.io').put('/').reply(200, 'success')
-
-    const result = await app.main({
+    const args = {
       token: 'abcdefg',
       url: 'https://codecov.io',
       parent,
       flags: '',
       slug: '',
       upstream: ''
-    })
+    }
+    const inputs = { args, environment: process.env }
+    const serviceParams = detectProvider(inputs, args.token != '')
+    const buildParams = webHelpers.populateBuildParams(inputs, serviceParams)
+    const query = webHelpers.generateQuery(buildParams)
+
+    mockClient = mockAgent.get('https://codecov.io')
+    mockClient.intercept({
+      method: 'POST',
+      path: `/upload/v4?package=uploader-${version}&token=${args.token}&${query}`,
+    }).reply(200, 'https://results.codecov.io\nhttps://codecov.io')
+
+    mockClient.intercept({
+      method: 'PUT',
+      path: '/',
+    }).reply(200, 'success')
+
+    const result = await app.main(args)
     expect(result).toEqual({
       status: 'success',
       resultURL: 'https://results.codecov.io/',